Renamed more icon
